Trannycamsundercumonmetranny@chaturbate.com (#1)
* Port to GNOME Shell 3.32.

* Update tr.po

Hi
Turkish translate updated

* Add Russian translation

* Add Russian language

* Update ru.po

* Update Makefile

* Support for gnome 3.22

* Minor updates

* Use ByteArray to avoid gnome-shell warning.

On gnome-shell 3.30 and 3.32, Netspeed emits this error message:
gnome-shell[25786]: Some code called array.toString() on a Uint8Array instance. Previously this would have interpreted the bytes of the array as a string, but that is nonstandard. In the future this will return the bytes as comma-separated digits. For the time being, the old behavior has been preserved, but please fix your code anyway to explicitly call ByteArray.toString(array).
                                           (Note that array.toString() may have been called implicitly.)
                                           0 _update() ["/home/XXXXXX/.local/share/gnome-shell/extensions/netspeed@hedayaty.gmail.com/net_speed.js":183]
                                           1 _update() ["self-hosted":975]

* Fix error in _reload().

_reload() can sometimes be called with _setting set to null.

* Make readme a markdown